+++ src/components/TopChartsCard.svelte
